flask_app/requirements.txt
pip~=25.0
pillow~=11.1.0
requests~=2.32.3
SQLAlchemy~=2.0.37
Flask~=3.1.0
typing_extensions~=4.12.2
MarkupSafe~=3.0.2
Werkzeug~=3.1.3
click~=8.1.8
Jinja2~=3.1.5
blinker~=1.9.0
itsdangerous~=2.2.0
psycopg2-binary~=2.9.10
pydantic~=2.10.6
pydantic_core~=2.27.2
annotated-types~=0.7.0
bcrypt~=4.2.1
python-dotenv~=1.0.1
Flask-Bcrypt~=1.0.1
pytest~=8.3.4
pluggy~=1.5.0
iniconfig~=2.0.0
certifi~=2025.1.31
urllib3~=2.3.0
greenlet~=3.1.1
psycopg2~=2.9.10
idna~=3.10
charset-normalizer~=3.4.1
packaging~=24.2